abstract parameters into trait
