+++ src/Omnipay/Common/AbstractGateway.php
@@ -42,8 +42,10 @@ use Omnipay\Common\Http\ClientInterface;
  *
  * @see GatewayInterface
  */
-abstract class AbstractGateway implements GatewayInterface
+abstract class AbstractGateway implements GatewayInterface, ParameterizedInterface
 {
+    use HasParametersTrait;
+
     /**
      * @var ParameterBag
      */
@@ -101,7 +103,7 @@ abstract class AbstractGateway implements GatewayInterface
             }
         }
 
-        Helper::initialize($this, $parameters);
+        Helper::initializeParameters($this, $parameters);
 
         return $this;
     }
@@ -112,35 +114,6 @@ abstract class AbstractGateway implements GatewayInterface
     public function getDefaultParameters()
     {
         return array();
-    }
-
-    /**
-     * @return array
-     */
-    public function getParameters()
-    {
-        return $this->parameters->all();
-    }
-
-    /**
-     * @param  string $key
-     * @return mixed
-     */
-    public function getParameter($key)
-    {
-        return $this->parameters->get($key);
-    }
-
-    /**
-     * @param  string $key
-     * @param  mixed  $value
-     * @return $this
-     */
-    public function setParameter($key, $value)
-    {
-        $this->parameters->set($key, $value);
-
-        return $this;
     }
 
     /**
