+++ src/Omnipay/Common/HasParametersTrait.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace Omnipay\Common;
+
+trait HasParametersTrait
+{
+    /**
+     * Internal storage of all of the parameters.
+     *
+     * @var ParameterBag
+     */
+    protected $parameters;
+
+    /**
+     * Set one parameter.
+     *
+     * @param string $key Parameter key
+     * @param mixed $value Parameter value
+     * @return $this
+     */
+    public function setParameter($key, $value)
+    {
+        $this->parameters->set($key, $value);
+
+        return $this;
+    }
+
+    /**
+     * Get one parameter.
+     *
+     * @return mixed A single parameter value.
+     */
+    public function getParameter($key)
+    {
+        return $this->parameters->get($key);
+    }
+
+    /**
+     * Get all parameters.
+     *
+     * @return array An associative array of parameters.
+     */
+    public function getParameters()
+    {
+        return $this->parameters->all();
+    }
+
+    /**
+     * Initialize the object with parameters.
+     *
+     * If any unknown parameters passed, they will be ignored.
+     *
+     * @param array $parameters An associative array of parameters
+     * @return $this.
+     */
+    public function initialize(array $parameters = [])
+    {
+        $this->parameters = new ParameterBag;
+
+        Helper::initializeParameters($this, $parameters);
+
+        return $this;
+    }
+}
