+++ tests/Omnipay/Common/HelperTest.php
@@ -46,13 +46,13 @@ class HelperTest extends TestCase
     public function testInitializeIgnoresNull()
     {
         $target = m::mock();
-        Helper::initialize($target, null);
+        Helper::initializeParameters($target, null);
     }
 
     public function testInitializeIgnoresString()
     {
         $target = m::mock();
-        Helper::initialize($target, 'invalid');
+        Helper::initializeParameters($target, 'invalid');
     }
 
     public function testInitializeCallsSetters()
@@ -61,7 +61,7 @@ class HelperTest extends TestCase
         $target->shouldReceive('setName')->once()->with('adrian');
         $target->shouldReceive('setNumber')->once()->with('1234');
 
-        Helper::initialize($target, array('name' => 'adrian', 'number' => '1234'));
+        Helper::initializeParameters($target, ['name' => 'adrian', 'number' => '1234']);
     }
 
     public function testInitializeIgnoresInvalidParameters()
@@ -69,7 +69,7 @@ class HelperTest extends TestCase
         $target = m::mock('\Omnipay\Common\CreditCard');
         $target->shouldReceive('setName')->once()->with('adrian');
 
-        Helper::initialize($target, array('name' => 'adrian', 'extra' => 'invalid'));
+        Helper::initializeParameters($target, ['name' => 'adrian', 'extra' => 'invalid']);
     }
 
     public function testGetGatewayShortNameSimple()
